% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_crossDynamicTable.R
\name{fct_crossDynamicTable}
\alias{fct_crossDynamicTable}
\title{fct_crossDynamicTable}
\usage{
fct_crossDynamicTable(
  tab,
  x_var,
  y_var = NULL,
  group_var = NULL,
  facet_x = NULL,
  facet_y = NULL,
  n_categ = 4,
  show_NA = FALSE
)
}
\arguments{
\item{tab}{Tableau de données source}

\item{x_var}{char : Variable en axe Y}

\item{y_var}{char : Variable en axe Y}

\item{group_var}{char : Variable de regroupement (color/fill selon le type de graphique)}

\item{facet_x}{char : Variable en facet x}

\item{facet_y}{char : Variable en facet y}

\item{n_categ}{int : Nombre de catégorie pour discrétiser group_var, facet_x et facet_y si elles sont continues}

\item{show_NA}{Booléen : Afficher ou non les données manquantes sur le graphique}
}
\value{
tibble
}
\description{
Réalise un tableau croisé dynamique correspondant au plot de fct_crossDynamicPlot selon les variables soumises
}
\examples{
library(EPAAVCshiny)
base = dplyr::as_tibble(mtcars) \%>\%
  dplyr::mutate(
    dplyr::across(c("cyl", "vs", "am", "gear", "carb"), ~factor(as.character(.)))
  )


base <- labelled::set_variable_labels(
  base,
  "mpg" = "Consommation (M/G)",
  "cyl" = "Nombre de cylindres",
  "disp" = "Displacement",
  "hp" = "Nombre de chevaux",
  "drat" = "Rear axle ratio",
  "wt" = "Poids",
  "qsec" = "1/4 mile time",
  "vs" = "Engine type",
  "am" = "Transmission type",
  "gear" = "Nombre de vitesses",
  "carb" = "Nombre de carburateurs")

fct_crossDynamicTable(tab = base, x_var = "mpg")
fct_crossDynamicTable(tab = base, x_var = "mpg", y_var = "am")
fct_crossDynamicTable(tab = base, x_var = "am", y_var = "carb")
fct_crossDynamicTable(tab = base, x_var = "mpg", y_var = "drat", group_var = "am")
fct_crossDynamicTable(tab = base, x_var = "am", y_var = "mpg", group_var = "gear")
fct_crossDynamicTable(tab = base, x_var = "am", y_var = "carb", group_var = "vs")
fct_crossDynamicTable(tab = base, x_var = "am", y_var = "carb", facet_x = "vs")
fct_crossDynamicTable(tab = base, x_var = "am", y_var = "carb", facet_y = "vs")
fct_crossDynamicTable(tab = base, x_var = "am", y_var = "carb", facet_x = "am", facet_y = "vs")
fct_crossDynamicTable(tab = base, x_var = "am", y_var = "carb",
                      facet_x = "am", facet_y = "wt", n_categ = 3)
}
