% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_utils.R
\name{plot_add_logo}
\alias{plot_add_logo}
\title{Add company logo to a plot}
\usage{
plot_add_logo(plot, height = 0.1, width = 0.2)
}
\arguments{
\item{plot}{A ggplot object}

\item{height}{The height of the logo in proportion of the graph (default 0.1)}

\item{width}{The width of the logo in proportion of the graph (default 0.2)}
}
\value{
a ggplot2 object
}
\description{
Adds Urg'Ara logo to a plot.
}
