% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tab_EPA_AVC}
\alias{tab_EPA_AVC}
\title{Ce jeu de données contient des informations issues de l'enquête EPA relatives aux accidents vasculaires cérébraux (AVC).}
\format{
Un data.frame avec X lignes et Y variables :
}
\usage{
tab_EPA_AVC
}
\description{
Ce jeu de données contient des informations issues de l'enquête EPA relatives aux accidents vasculaires cérébraux (AVC).
}
\keyword{datasets}
