% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_utils.R
\name{utils_get_label}
\alias{utils_get_label}
\title{Extract label variable}
\usage{
utils_get_label(tab, var)
}
\arguments{
\item{tab}{tableau de données dans lequel se trouve la variable}

\item{var}{Char : Nom de la variable}
}
\value{
char
}
\description{
Extrait le champ label d'une variable. Retourne une erreur si pas de label
}
