% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_utils.R
\name{utils_get_scale_type}
\alias{utils_get_scale_type}
\title{Get scale type}
\usage{
utils_get_scale_type(var, axis)
}
\arguments{
\item{var}{vecteur variable à tester}

\item{axis}{axe (x/y) dans lequel la variable sera utilisée}
}
\value{
char
}
\description{
Détermine le type de scale à appliquer à un ggplot selon le type d'une variable
}
