% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_utils.R
\name{utils_is.continuous}
\alias{utils_is.continuous}
\title{check continuous
Vérifie qu'une variable est numéric (double/int) ou temporelle}
\usage{
utils_is.continuous(var)
}
\arguments{
\item{var}{vecteur variable a tester}
}
\value{
Booléen
}
\description{
check continuous
Vérifie qu'une variable est numéric (double/int) ou temporelle
}
