% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_utils.R
\name{utils_is.discrete}
\alias{utils_is.discrete}
\title{check discrete}
\usage{
utils_is.discrete(var)
}
\arguments{
\item{var}{vecteur variable a tester}
}
\value{
Booléen
}
\description{
Vérifie qu'une variable est character ou factor
}
